Commit Inicial

+++ src/main/java/br/com/rotafood/api/application/dto/logistic/VrpInDto.java
@@ -1,5 +1,6 @@
 package br.com.rotafood.api.application.dto.logistic;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
@@ -9,7 +10,7 @@ public record VrpInDto (
     UUID id,
     VrpOriginDto origin,
     List<VrpOrderDto> orders, 
-    float maxRouteVolume,
+    BigDecimal maxRouteVolume,
     Date createdAt
 ) {
